Add basic types and trait implementations for fs::Tree
I switched the tree back to cloning its items. This allows users of the
tree to wrap items in an Arc if they want to, which makes it easier to
utilize the fact that items are heap allocated elsewhere in the system.
In this case, I wanted to avoid cloning entries in order to place them
within relative paths.

I use a SmallVec of size 1 because the relative paths summarizing
individual items will always be 1 element long, which makes a heap
allocation in that case wasteful. I don't want to reserve space for more
than one element however, because that would end up bloating the size of
the tree.